% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries}
\alias{countries}
\title{Country Codes}
\format{Character vector of length 182}
\source{
Feenstra, Robert C., Robert Inklaar and Marcel P. Timmer (2015), "The Next Generation of the Penn World Table", American Economic Review, 105(10), 3150-3182, available for download at \url{https://www.rug.nl/ggdc/productivity/pwt/}
}
\usage{
countries
}
\description{
A vector containing ISO country codes used in the Penn World Table.
}
\keyword{datasets}
