% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country3}
\alias{country3}
\title{Country Codes}
\format{Character vector of length 249}
\source{
\url{https://en.wikipedia.org/wiki/ISO_3166}
}
\usage{
country3
}
\description{
A vector containing ISO 3-digit country codes.
}
\keyword{datasets}
