% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_updates.R
\name{fred_updates}
\alias{fred_updates}
\title{Fetch Series with Recent Updates}
\usage{
fred_updates(key = NULL, args = list())
}
\arguments{
\item{key}{32 character lower-cased alpha-numeric character string}

\item{args}{named list of other arguments passed to the API. See details.}
}
\value{
data.frame with values for the series id, its title, start and end date of its observations, its frequency, units, whether it is seasonally adjusted, its popularity, when it was last updated, and the realtime start and end dates
}
\description{
Function that returns a data.frame containing information on series in FRED that were recently updated.
}
\details{
Arguments accept any of the following:

\describe{
 \item{\code{realtime_start}, \code{realtime_end}}{character string of format \code{"YYYY-MM-DD"} determining the real-time period. This is used to find historical series.}
 \item{\code{limit}}{integer between 1 and 1000 determining the maximum number of results to return.}
 \item{\code{offset}}{Non-negative integer}
 \item{\code{filter_value}}{Either \code{"macro"}, \code{"regional"}, or \code{"all"} *(the default)* describing the geographic type of the series}
 \item{\code{start_time}, \code{end_time}}{character string of format \code{"YYYYMMDDHhmm"} determining the time range, down to minutes, with which to filter the results}
}
}
