% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_search.R
\name{parse_search}
\alias{parse_search}
\title{Convert Search Terms to a API Friendly Format}
\usage{
parse_search(...)
}
\arguments{
\item{...}{character vectors of search terms}
}
\value{
character string
}
\description{
The function replaces spaces from character vectors with \code{"+"} and separates
search terms with \code{";"}. Use whenever search terms are needed in the \code{args}
list for API calls. It is automatically applied for \code{\link{fred_search}} and not needed.
}
\examples{
terms <- c("Real GDP", "medical inflation", "Japan")
parse_search(terms, "unemployment")
}
