% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{releases}
\alias{releases}
\title{Economic Releases}
\format{data.frame with 292 observations and 7 variables:
\describe{
  \item{id}{release id}
  \item{realtime_start, realtime_end}{date releases were saved}
  \item{name}{name of the release}
  \item{press_release}{logical value for whether there is an associated press release}
  \item{link}{a url for more data}
  \item{notes}{character string containing information about the release}
}}
\source{
\url{https://fred.stlouisfed.org/docs/api/fred/releases.html}
}
\usage{
releases
}
\description{
A data.frame containing information on the economic releases in FRED. Information
is from January 2021. Use \code{\link{fred_releases}} for an up-to-date version.
}
\keyword{datasets}
