% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{set_fred_key}
\alias{set_fred_key}
\title{Set a FRED API Key}
\usage{
set_fred_key(key)
}
\arguments{
\item{key}{32 character lower-cased alpha-numeric character string}
}
\value{
\code{NULL}
}
\description{
Function used to set the default key. If included, then \code{key} arguments
in FRED API calls are no longer needed. API keys can be requested at
\url{https://fred.stlouisfed.org/docs/api/api_key.html}.
}
\examples{
set_fred_key("abcdefghijklmnopqrstuvwxyz123456")

}
