% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sources}
\alias{sources}
\title{Sources of Economic Data}
\format{data.frame with 102 observations and 6 variables:
\describe{
  \item{id}{source id}
  \item{realtime_start, realtime_end}{date releases were saved}
  \item{name}{name of the source}
  \item{link}{a url that links to the source's website}
  \item{notes}{character string containing information about the source}
}}
\source{
\url{https://fred.stlouisfed.org/docs/api/fred/sources.html}
}
\usage{
sources
}
\description{
A data.frame containing information on the sources of data in FRED. Information
is from January 2021. Use \code{\link{fred_sources}} for an up-to-date version.
}
\keyword{datasets}
