% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states}
\alias{states}
\title{U.S. State Abbreviations}
\format{Character vector of length 51}
\source{
\url{https://en.wikipedia.org/wiki/List_of_U.S._state_and_territory_abbreviations}
}
\usage{
states
}
\description{
A vector containing U.S. state abbreviations, including the District of Columbia.
}
\keyword{datasets}
