% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states_extended}
\alias{states_extended}
\title{Extended State Abbreviations}
\format{Character vector of length 51}
\source{
\url{https://en.wikipedia.org/wiki/List_of_U.S._state_and_territory_abbreviations}
}
\usage{
states_extended
}
\description{
A vector containing abbreviations of U.S. states and territories.
}
\keyword{datasets}
