% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tags}
\alias{tags}
\title{Popular Tags}
\format{data.frame with 1000 observations and 6 variables:
\describe{
  \item{name}{tag name}
  \item{group_id}{name of the group associated with the tag}
  \item{created}{when the tag was created}
  \item{popularity}{integer of the tag's popularity (up to 100)}
  \item{notes}{character string containing information about the source}
  \item{series_count}{number of series associated with the tag}
}}
\source{
\url{https://fred.stlouisfed.org/docs/api/fred/tags.html}
}
\usage{
tags
}
\description{
A data.frame containing information on 1000 of the most common tags in FRED. Information
is from January 2021. Use \code{\link{fred_tags}} for an up-to-date version
and the ability to refine it further.
}
\keyword{datasets}
